# BUG: Fix dataframe.update not updating dtype #55509 v2