--- a/pandas/tests/frame/methods/test_update.py
+++ b/pandas/tests/frame/methods/test_update.py
@@ -184,3 +184,32 @@ class TestDataFrameUpdate:
             {"A": [1.0, 3.0], "B": [pd.NaT, pd.to_datetime("2016-01-01")]}
         )
         tm.assert_frame_equal(df, expected)
+
+    @pytest.mark.parametrize(
+        "value_df, value_other, dtype",
+        [
+            (True, False, bool),
+            (1, 2, int),
+            (np.uint64(1), np.uint(2), np.dtype("uint64")),
+            (1.0, 2.0, float),
+            (1.0 + 1j, 2.0 + 2j, complex),
+            ("a", "b", pd.StringDtype()),
+            (
+                pd.to_timedelta("1 ms"),
+                pd.to_timedelta("2 ms"),
+                np.dtype("timedelta64[ns]"),
+            ),
+            (
+                np.datetime64("2000-01-01T00:00:00"),
+                np.datetime64("2000-01-02T00:00:00"),
+                np.dtype("datetime64[ns]"),
+            ),
+        ],
+    )
+    def test_update_preserve_dtype(self, value_df, value_other, dtype):
+        # GH#55509
+        df = DataFrame({"a": [value_df] * 2}, index=[1, 2])
+        other = DataFrame({"a": [value_other]}, index=[1])
+        expected = DataFrame({"a": [value_other, value_df]}, index=[1, 2])
+        df.update(other)
+        tm.assert_frame_equal(df, expected)