@aureliobarbosa 2024-03-01
@datapythonista 

> Do you think it's worth adding a use case to test when the indices have no intersection?

I agree because this is a behavior that was introduced on this PR. Added the test for this case and changed the message raise slightly to emphasize 'no intersection'.

> not sure if we can find a better name for the variable rows. Maybe mask or index_intersection is better.

The first of this PR used `indexes_intersection`, but then I changed to `rows` to simplify the notation. Now that you raised this point, I noted that a simple search on the file `frame.py` results in more than one hundred matches. Considering this, I agree with avoided using `rows`.