@aureliobarbosa 2024-03-02
I did benchmark again the first three manual code tests proposed previously by @rhshadrach. Results are summarized below in the same order.

| Data | main  | PR  | 
| --- | --- | --- |
| monotonic  |  1.8 ms  | 1.29  ms  |  
| non-monotonic |  1.83 ms | 1.21 ms  |  
| small-frame-large-arg | 384 µs |595 µs | 

The performance gain is not as good as originally expected, but they are still relevant. @datapythonista Note that when the frame is small and the argument is large there is a performance loss, but as pointed out previously by @rhshadrach  this looks like an edge use case.