@aureliobarbosa 2024-03-02
Hope to be done here. 

@datapythonista Next thing I will do is to try to implement an ASV benchmark for this function.

@rhshadrach I think that the take home message on this PR was that the function was manipulating data it wasn't expecting to touch on, and this case, the change in the dtype was the symptom for this problem.  

If you both know or find any bug that could be related or similar to this one, I would be glad to contribute.

Regards